requirements.txt
lightning
pandas
matplotlib
seaborn
scikit-learn
python-dotenv
wandb
Faker
kaggle
